Add flags to make compilation of various solvers optional
This is mainly for SciPy since they would like to trim the binary size
by not including the other solvers. By default all solvers will be
built, so the way to use this is to define DONT_BUILD_BOBYQA or whatever
algorithm you don't want.

The use of the double negative is unfortunate, but the alternative would
be to require users to explicitly build the algorithms they want, which
seems burdensome.