Add comparing facilities.
In order to better look for bugs in the implementation as compared to
the Fortran, I've added tooling to more accurately compare floating
point operations. Fortran code needs to be compiled in debug mode for
these to make sense, and they make the code much much slower, but it can
be helpful for detecting issues in the translation. Unfortunately, even
with these changes, we still do not achieve bit for bit precision
compared to Fortran, but at this point the search for bit for bit
precision is not yielding much fruit while consuming much effort.