Initial commit consolidating python and pyprima
This is just a quick proof of concept to show how the code consolidation
will work. I tested this by building a wheel for the python bindings and
then testing it in an isolated virtual environment and it worked as
expected. I also checked the folder structure of the 'prima' package
that was installed via the wheel and could see that it correctly copied
pycommon as expected.

This is in support of #252.