Move process_linear_constraints into pycommon
This one necessitated creating a symlink to the objective function, and
also dealing with the fact that the python bindings return an object
with 'fun' as an attribute, whereas pyprima more closely follows the
Fortran and returns an object with 'f' as an attribute for the function
value.

We should replace 'f' with 'fun' in pyprima to be consistent with scipy,
the python bindings, and COBYQA (which grabs OptimizeResult from scipy
and is therefore always consistent with it).

Ref: #252