Move process_constraints to pycommon
This is a somewhat substantial move as compared to the others, but it's
a significant step in the consolidation.

This move revealed that it was a mistake to check the "constraints"
against "Iterable" in process_constraints, as a dictionary counts as
Iterable, but we would still want to wrap it in an array. I'm not sure
why pyprima used Iterable, hopefully this does not cause any issues
going forward.

Ref: #252