Rearrange python files to keep them in one subfolder
pyproject.toml needs to stay at the root level because python can only
create source distributions at the root level. Attempts were made to
have pyproject.toml not at the root level, but the only potential
solutions introduced significant levels of complexity with potential
maintenance issues, and so it was decided that keeping it at the root
level is the simplest and best solution.